Make COpenGLSync public, allowing for Interoperation with SDKs also using the OpenGL API
